# api-spec/openapi/pilot-v1.yaml
...
openapi: 3.1.0
...
info:
  title: Pilot API
  version: 1.0.0
  contact:
    name: bcodesphere
  description: >-
    API REST de Pilot, ERP multiempresa para PYMES de El Salvador. Incluye el núcleo
    (usuarios y empresas) y la app Contabilidad. Los montos viajan como cadena decimal
    con 2 decimales (ADR-013) y los errores siguen RFC 9457 (Problem Details).

# --- Servidores: ruta base relativa al host que sirve la API ---
servers:
  - url: /api/v1
    description: Ruta base de la versión 1 de la API.

# --- Etiquetas por módulo del backend (CLAUDE.md §4.2) ---
...
security:
  - oidc: []
paths:
...
  /contabilidad/cuentas:
    get:
      operationId: listarCuentasContables
      tags: [cuentasContables]
      x-rol-minimo: auditor
      summary: Lista el catálogo completo de cuentas
      description: >-
        Devuelve el catálogo de cuentas de la empresa completo y sin paginar, ordenado por código, para armar el árbol y
        la búsqueda en el cliente (ADR-035). Los filtros son opcionales y se combinan: busqueda (código o nombre, sin
        distinguir mayúsculas), soloDetalle (solo cuentas que aceptan movimientos) y soloActivas. Rol mínimo: auditor.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
        - $ref: '#/components/parameters/BusquedaCuenta'
        - $ref: '#/components/parameters/SoloDetalle'
        - $ref: '#/components/parameters/SoloActivas'
      responses:
        '200':
          description: Catálogo de cuentas ordenado por código.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CuentaContable'
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
          description: "400 — PLT-001: parámetros con tipos incorrectos."
        '422':
          $ref: '#/components/responses/EntidadNoProcesable'
          description: "422 — PLT-002: validación de forma de los parámetros (incluye errores)."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
    post:
      operationId: crearCuentaContable
      tags: [cuentasContables]
      x-rol-minimo: contador
      summary: Crea una cuenta contable
      description: >-
        Crea una cuenta del catálogo. La cuenta padre no se envía: se deduce del código, porque el código del padre
        es prefijo del de la hija (ADR-035). La clase es el primer dígito del código; la naturaleza no se envía,
        siempre se deriva de la cuenta padre, o de la clase en el nivel 1 (deudora en 1 y 4, acreedora en 2, 3 y 5,
        ADR-042). Rol mínimo: contador.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
      requestBody:
        required: true
        description: Código y nombre de la nueva cuenta.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuevaCuentaContable'
      responses:
        '201':
          description: Cuenta creada con su ETag de versión.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaContable'
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
          description: "400 — PLT-001: cuerpo JSON ilegible o con tipos incorrectos."
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '409':
          $ref: '#/components/responses/Conflicto'
          description: "409 — CON-014: el código ya existe en el catálogo de la empresa."
        '422':
          $ref: '#/components/responses/EntidadNoProcesable'
          description: "422 — CON-010: el primer dígito del código no es una clase de 1 a 5; CON-011: la cuenta padre tiene movimientos; CON-015: longitud de código no válida (1, 2, 4, 6 u 8 dígitos) o sin cuenta padre existente y activa cuyo código sea su prefijo; PLT-002: validación de forma (incluye errores)."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /contabilidad/cuentas/{cuentaId}:
    get:
      operationId: obtenerCuentaContable
      tags: [cuentasContables]
      x-rol-minimo: auditor
      summary: Consulta una cuenta contable
      description: >-
        Devuelve una cuenta del catálogo con el header ETag de su versión actual, que debe enviarse como If-Match al
        editarla. Rol mínimo: auditor.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
        - $ref: '#/components/parameters/CuentaIdRuta'
      responses:
        '200':
          description: Cuenta con su ETag de versión.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaContable'
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '404':
          $ref: '#/components/responses/NoEncontrado'
          description: "404 — PLT-017: el recurso no existe o no pertenece a la empresa activa."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
    patch:
      operationId: actualizarCuentaContable
      tags: [cuentasContables]
      x-rol-minimo: contador
      summary: Edita una cuenta contable
      description: >-
        Cambia el código, el nombre o el estado activa de una cuenta; solo se envían los campos a cambiar. La
        naturaleza no se edita: es de solo lectura, derivada de la cuenta padre o de la clase (ADR-042). Exige
        If-Match con el ETag leído (concurrencia optimista) y devuelve el nuevo ETag. El código no puede cambiarse
        si la cuenta tiene movimientos (CON-011) y una cuenta con saldo distinto de cero no puede desactivarse
        (CON-012). Si la cuenta está en uso por la configuración contable o por una regla activa, no se puede
        desactivar ni dejar de ser de detalle (CON-016). Una cuenta del sistema (catálogo base) no se edita
        (CON-021, ADR-042). Rol mínimo: contador.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
        - $ref: '#/components/parameters/CuentaIdRuta'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        description: Campos de la cuenta a modificar (al menos uno).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizacionCuentaContable'
      responses:
        '200':
          description: Cuenta actualizada con su nuevo ETag.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaContable'
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
          description: "400 — PLT-001: cuerpo JSON ilegible o con tipos incorrectos."
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '404':
          $ref: '#/components/responses/NoEncontrado'
          description: "404 — PLT-017: el recurso no existe o no pertenece a la empresa activa."
        '409':
          $ref: '#/components/responses/Conflicto'
          description: "409 — CON-014: el nuevo código ya existe en el catálogo de la empresa."
        '412':
          $ref: '#/components/responses/PrecondicionFallida'
          description: "412 — PLT-016: If-Match no coincide con la versión actual; volver a leer el recurso."
        '422':
          $ref: '#/components/responses/EntidadNoProcesable'
          description: "422 — CON-010: el primer dígito del código no es una clase de 1 a 5; CON-011: cambiar el código de una cuenta con movimientos, o crear una hija en una cuenta con movimientos; CON-012: desactivar una cuenta con saldo distinto de cero; CON-015: longitud de código no válida o sin cuenta padre existente y activa cuyo código sea su prefijo; CON-016: la cuenta está en uso por la configuración contable o por una regla activa; CON-021: la cuenta es del sistema y no se puede modificar; PLT-002: validación de forma (incluye errores)."
        '428':
          $ref: '#/components/responses/PrecondicionRequerida'
          description: "428 — PLT-015: falta el header If-Match."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'

  # --- Configuración contable (CLAUDE.md §11.3): una por empresa, PUT completo ---
...
components:
...
  securitySchemes:
    oidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso JWT emitido por Keycloak (OIDC con PKCE) para usuarios de la web.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: pk_xxxx.secreto
      description: API key de integración con formato pk_xxxx.secreto, usada por n8n; la empresa se deriva de la clave.

  # --- Esquemas reutilizables ---
  schemas:
...
    MontoConSigno:
      type: string
      pattern: '^-?\d{1,17}(\.\d{1,2})?$'
      examples: ['-10.50']
      description: >-
        Monto en USD que puede ser negativo (p. ej. una diferencia), como cadena decimal con máximo
        2 decimales (ADR-013). El servidor siempre responde con exactamente 2 decimales.
    Uuid:
      type: string
      format: uuid
      description: Identificador UUID (versión 7 generado por la aplicación, ADR-010).
...
    ErrorCampo:
      type: object
      description: Detalle de un error de validación asociado a un campo.
      required: [campo, mensaje]
      properties:
        campo:
          type: string
          description: Ruta del campo con error dentro del cuerpo o de los parámetros.
        mensaje:
          type: string
          description: Mensaje de error en español.
    ProblemDetails:
      type: object
      description: Error según RFC 9457 (application/problem+json) con extensiones de Pilot.
      required: [type, title, status, codigo]
      properties:
        type:
          type: string
          description: URI que identifica el tipo de problema; about:blank si no hay uno específico.
        title:
          type: string
          description: Resumen corto del problema, en español.
        status:
          type: integer
          description: Código de estado HTTP de la respuesta.
        detail:
          type: string
          description: Explicación del problema para esta ocurrencia; sin trazas ni datos internos.
        instance:
          type: string
          description: URI o identificador de la ocurrencia concreta del problema.
        codigo:
          type: string
          pattern: '^(PLT|CON|INT)-\d{3}$'
          description: Código de negocio con prefijo del módulo (PLT plataforma, CON contabilidad, INT integración).
        errores:
          type: array
          description: Errores de validación por campo; solo en fallos de validación.
          items:
            $ref: '#/components/schemas/ErrorCampo'
        diferencia:
          $ref: '#/components/schemas/MontoConSigno'
          description: Diferencia Σ Debe − Σ Haber; solo en descuadres y cobros que no cuadran.
        operacionId:
          $ref: '#/components/schemas/Uuid'
          description: Operación externa ya contabilizada; solo en 409 INT-004 (ADR-039 §5).
        asientoId:
          $ref: '#/components/schemas/Uuid'
          description: Asiento generado por la operación ya contabilizada; solo en 409 INT-004 (ADR-039 §5).
...
    NaturalezaCuenta:
      type: string
      enum: [DEUDORA, ACREEDORA]
      description: Naturaleza de la cuenta; deudora aumenta con el Debe (clases 1 y 4), acreedora con el Haber (2, 3 y 5).
...
    CuentaContable:
      type: object
      description: Cuenta del catálogo de la empresa; también viaja su versión como ETag.
      required:
        [id, codigo, nombre, clase, nivel, cuentaPadreId, naturaleza, aceptaMovimientos, activa, version, sistema]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        codigo:
          type: string
          description: Código de la cuenta; su primer dígito es la clase.
        nombre:
          type: string
          description: Nombre de la cuenta.
        clase:
          type: integer
          minimum: 1
          maximum: 5
          description: Clase de la cuenta (primer dígito del código) 1 Activo, 2 Pasivo, 3 Capital, 4 Costos y Gastos, 5 Ingresos.
        nivel:
          type: integer
          minimum: 1
          maximum: 5
          description: Nivel por longitud del código, 1 clase, 2 grupo, 3 cuenta, 4 subcuenta, 5 detalle.
        cuentaPadreId:
          type: [string, 'null']
          format: uuid
          description: Cuenta padre (su código es prefijo del de esta cuenta); nulo en las clases.
        naturaleza:
          $ref: '#/components/schemas/NaturalezaCuenta'
          description: >-
            Naturaleza de la cuenta; de solo lectura, derivada de la cuenta padre o de la clase en el nivel 1
            (ADR-042). Ya no se envía al crear ni al editar una cuenta.
        aceptaMovimientos:
          type: boolean
          description: true solo en las cuentas de detalle, sin hijas; únicas que aceptan movimientos.
        activa:
          type: boolean
          description: Si la cuenta está activa; una cuenta inactiva no acepta movimientos.
        version:
          type: integer
          format: int64
          description: Versión para concurrencia optimista; también viaja como ETag (CLAUDE.md §8.3).
        sistema:
          type: boolean
          readOnly: true
          description: >-
            true si la cuenta viene del catálogo base: su código, nombre, naturaleza y estado no se editan
            (422 CON-021, ADR-042). El usuario agrega subcuentas propias bajo las cuentas base que no son de
            detalle, y esas sí las edita.
    NuevaCuentaContable:
      type: object
      description: >-
        Datos para crear una cuenta. La cuenta padre no se envía porque se deduce del código (ADR-035). El patrón
        del código no exige la clase para que la regla de negocio responda CON-010 y CON-015. La naturaleza no se
        envía: siempre se deriva de la cuenta padre, o de la clase en el nivel 1 (ADR-042).
      required: [codigo, nombre]
      properties:
        codigo:
          type: string
          pattern: '^[0-9]{1,8}$'
          description: Código de 1 a 8 dígitos; su primer dígito debe ser una clase de 1 a 5 (CON-010).
        nombre:
          type: string
          minLength: 1
          maxLength: 200
          description: Nombre de la cuenta.
    ActualizacionCuentaContable:
      type: object
      description: >-
        Campos editables de una cuenta; solo se envían los que cambian y debe venir al menos uno. La naturaleza no
        se envía: es de solo lectura, derivada de la cuenta padre o de la clase (ADR-042).
      minProperties: 1
      properties:
        codigo:
          type: string
          pattern: '^[0-9]{1,8}$'
          description: Nuevo código; no puede cambiarse si la cuenta tiene movimientos (CON-011).
        nombre:
          type: string
          minLength: 1
          maxLength: 200
          description: Nuevo nombre de la cuenta.
        activa:
          type: boolean
          description: false desactiva la cuenta (CON-012 si tiene saldo, CON-016 si está en uso); true la reactiva.

    # --- Configuración contable (CLAUDE.md §11.3) ---
...
  parameters:
    XEmpresaId:
      name: X-Empresa-Id
      in: header
      required: true
      description: Empresa activa de la sesión; se valida contra las membresías del usuario (CLAUDE.md §4.5).
      schema:
        $ref: '#/components/schemas/Uuid'
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Identificador de correlación; se genera si no viene y se devuelve en la respuesta.
      schema:
        type: string
        maxLength: 100
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag de la versión leída del recurso; si no coincide con la actual se responde 412 PLT-016 (CLAUDE.md §8.3).
      schema:
        type: string
        minLength: 1
...
    CuentaIdRuta:
      name: cuentaId
      in: path
      required: true
      description: Identificador de la cuenta contable.
      schema:
        $ref: '#/components/schemas/Uuid'
...
    BusquedaCuenta:
      name: busqueda
      in: query
      required: false
      description: Texto a buscar en el código o el nombre de la cuenta, sin distinguir mayúsculas.
      schema:
        type: string
        maxLength: 100
    SoloDetalle:
      name: soloDetalle
      in: query
      required: false
      description: true devuelve solo las cuentas de detalle (las que aceptan movimientos).
      schema:
        type: boolean
        default: false
    SoloActivas:
      name: soloActivas
      in: query
      required: false
      description: true omite las cuentas inactivas.
      schema:
        type: boolean
        default: false
...
  headers:
    X-Request-Id:
      description: Identificador de correlación de la petición, para soporte y trazas.
      schema:
        type: string
    ETag:
      description: Versión del recurso (su columna version) entre comillas; se envía como If-Match al editarlo.
      schema:
        type: string
...
    Retry-After:
      description: Segundos que el cliente debe esperar antes de reintentar.
      schema:
        type: integer
...
  responses:
    SolicitudInvalida:
      description: 400 — El cuerpo o los parámetros no cumplen el contrato (incluye la lista errores).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NoAutenticado:
      description: 401 — Falta la credencial o es inválida, está vencida o fue revocada.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Prohibido:
      description: 403 — La credencial es válida pero no tiene rol, alcance o app activa para esta operación.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NoEncontrado:
      description: 404 — El recurso no existe o no pertenece a la empresa activa.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflicto:
      description: 409 — El estado actual del recurso impide la operación (duplicado o petición en proceso).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    PrecondicionFallida:
      description: 412 — El header If-Match no coincide con la versión actual del recurso; volver a leerlo.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    EntidadNoProcesable:
      description: 422 — La petición es válida en forma pero viola una regla de negocio (incluye diferencia si aplica).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    PrecondicionRequerida:
      description: 428 — Falta un header obligatorio de precondición, como Idempotency-Key o If-Match.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    DemasiadasPeticiones:
      description: 429 — Se superó el límite de peticiones; ver el header Retry-After.
      headers:
        Retry-After:
          $ref: '#/components/headers/Retry-After'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ErrorInterno:
      description: 500 — Error interno; no expone trazas ni mensajes internos. Reintentar es seguro con idempotencia.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'